Create DetektVanilla to make it similar to how DetektJvm and DetektAndroid are registered
